# Conflicts:
#	BitPay/AccessTokens.cs
#	BitPay/BitPay.cs
#	BitPay/Client.cs
#	BitPay/Clients/AuthorizationClient.cs
#	BitPay/Clients/BillClient.cs
#	BitPay/Clients/BitPayClient.cs
#	BitPay/Clients/InvoiceClient.cs
#	BitPay/Clients/LedgerClient.cs
#	BitPay/Clients/PayoutClient.cs
#	BitPay/Clients/PayoutRecipientsClient.cs
#	BitPay/Clients/RateClient.cs
#	BitPay/Clients/RefundClient.cs
#	BitPay/Clients/SettlementClient.cs
#	BitPay/Clients/WalletClient.cs
#	BitPay/Models/Currency.cs
#	BitPay/Models/Invoice/Refund.cs
#	BitPay/Models/Payout/Payout.cs
#	BitPay/Models/Payout/PayoutBatch.cs
#	BitPayUnitTest/BitPayTest.cs
#	BitPayXUnitTest/BitPayTests.cs

+++ BitPay/Models/Invoice/InvoiceTransaction.cs
@@ -10,7 +10,7 @@ namespace BitPay.Models.Invoice
         private dynamic _exchangeRates;
         
         [JsonProperty(PropertyName = "amount")]
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
 
         [JsonProperty(PropertyName = "confirmations")]
         public string Confirmations { get; set; }
@@ -25,7 +25,7 @@ namespace BitPay.Models.Invoice
         public string Txid { get; set; }
         
         [JsonProperty(PropertyName = "refundAmount")]
-        public double RefundAmount { get; set; }
+        public decimal RefundAmount { get; set; }
 
         [JsonProperty(PropertyName = "time")]
         public string Time { get; set; }
